auth function code refactor and reusable (#305)
* rearrange some auth funcs

* clean up imports

* use ego package utils

* add rdpc code from env var

* remove rdpc_code property from base query args

---------
