+++ .schema.env
@@ -64,3 +64,9 @@ PORT=7000
 ######
 GRAPHQL_PORT=
 
+
+######
+# RDPC
+######
+RDPC_CODE=TORONTO
+
